# nodejs-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is the API for the Estuary application."
  version: "0.0.0"
  title: "Estuary API"
  termsOfService: "http://estuary.tech"
  contact:
    name: "API Support"
    url: "https://docs.estuary.tech/feedback"
  license:
    name: "Apache 2.0 Apache-2.0 OR MIT"
    url: "https://github.com/application-research/estuary/blob/master/LICENSE.md"
host: "api.estuary.tech"
basePath: "/"
security:
- bearerAuth: []
paths:
  /admin/autoretrieve/init:
    post:
      tags:
      - "autoretrieve"
      summary: "Register autoretrieve server"
      description: "This endpoint registers a new autoretrieve server Q@&^5#^&*Q^Wasdfalsdkjflkj"
      operationId: "adminAutoretrieveInitPOST"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "addresses"
        description: "Autoretrieve's comma-separated list of addresses"
        required: true
        schema:
          type: "string"
      - in: "body"
        name: "pubKey"
        description: "Autoretrieve's public key"
        required: true
        schema:
          type: "string"
      responses: {}
      x-swagger-router-controller: "Autoretrieve"
...
  /autoretrieve/heartbeat:
    post:
      tags:
      - "autoretrieve"
      summary: "Marks autoretrieve server as up"
      description: "This endpoint updates the lastConnection field for autoretrieve"
      operationId: "autoretrieveHeartbeatPOST"
      produces:
      - "application/json"
      parameters:
      - name: "token"
        in: "header"
        description: "Autoretrieve's auth token"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Autoretrieve"
...
  /deal/estimate:
    post:
      tags:
      - "deals"
      summary: "Estimate the cost of a deal"
      description: "This endpoint estimates the cost of a deal"
      operationId: "dealEstimatePOST"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "The size of the deal in bytes, the replication factor, and the\
          \ duration of the deal in blocks"
        required: true
        schema:
          $ref: "#/definitions/main.estimateDealBody"
      responses: {}
      x-swagger-router-controller: "Deals"
  /deal/info/{dealid}:
    get:
      tags:
      - "deals"
      summary: "Get Deal Info"
      description: "This endpoint returns the deal info for a deal"
      operationId: "dealInfoDealidGET"
      produces:
      - "application/json"
      parameters:
      - name: "dealid"
        in: "path"
        description: "Deal ID"
        required: true
        type: "integer"
      responses: {}
      x-swagger-router-controller: "Deals"
  /deal/proposal/{propcid}:
    get:
      tags:
      - "deals"
      summary: "Get Proposal"
      description: "This endpoint returns the proposal for a deal"
      operationId: "dealProposalPropcidGET"
      produces:
      - "application/json"
      parameters:
      - name: "propcid"
        in: "path"
        description: "Proposal CID"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Deals"
  /deal/query/{miner}:
    get:
      tags:
      - "deals"
      summary: "Query Ask"
      description: "This endpoint returns the ask for a given CID"
      operationId: "dealQueryMinerGET"
      produces:
      - "application/json"
      parameters:
      - name: "miner"
        in: "path"
        description: "CID"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Deals"
  /deal/status-by-proposal/{propcid}:
    get:
      tags:
      - "deals"
      summary: "Get Deal Status by PropCid"
      description: "Get Deal Status by PropCid"
      operationId: "dealStatus_by_proposalPropcidGET"
      produces:
      - "application/json"
      parameters:
      - name: "propcid"
        in: "path"
        description: "PropCid"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Deals"
  /deal/status/{miner}/{propcid}:
    get:
      tags:
      - "deals"
      summary: "Deal Status"
      description: "This endpoint returns the status of a deal"
      operationId: "dealStatusMinerPropcidGET"
      produces:
      - "application/json"
      parameters:
      - name: "miner"
        in: "path"
        description: "Miner"
        required: true
        type: "string"
      - name: "propcid"
        in: "path"
        description: "Proposal CID"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Deals"
...
securityDefinitions:
  bearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  main.estimateDealBody:
    type: "object"
    properties:
      durationBlks:
        type: "integer"
      replication:
        type: "integer"
      size:
        type: "integer"
      verified:
        type: "boolean"
...